refactor(app): extract useJobHistoryPanel from ManageAnnotations (#399)
Reduce ManageAnnotations.vue from 690 to 614 lines by bundling the
job-history panel concern into a useJobHistoryPanel composable: row data,
search-filtered + paginated derivations, CSV download, page-link copy, and
clear-all-filters, composing the existing useJobHistoryUrlState for the
URL-synced table controls.

Behavior, network calls, and the JobHistoryCard contract are unchanged.

Advances #399. Part of #346.
